Initial refactoring of websockets
