+++ packages/webapi/websocket/events.go
@@ -1,11 +1,14 @@
-package publisher
+package websocket
 
 import (
 	"fmt"
 
+	"github.com/iotaledger/hive.go/core/generics/event"
 	"github.com/iotaledger/hive.go/core/logger"
+	"github.com/iotaledger/hive.go/core/websockethub"
 	"github.com/iotaledger/wasp/packages/kv"
 	"github.com/iotaledger/wasp/packages/kv/subrealm"
+	"github.com/iotaledger/wasp/packages/publisher"
 	"github.com/iotaledger/wasp/packages/vm/core/blocklog"
 )
 
@@ -13,16 +16,15 @@ const (
 	ISCEventKindNewBlock      = "new_block"
 	ISCEventKindReceipt       = "receipt" // issuer will be the request sender
 	ISCEventKindSmartContract = "contract"
+	ISCEventIssuerVM          = "vm"
 )
-
-const ISCEventIssuerVM = "vm"
 
 type ISCEvent struct {
 	Kind      string
 	Issuer    string // (AgentID) nil means issued by the VM
 	RequestID string // (isc.RequestID)
 	ChainID   string // (isc.ChainID)
-	Content   interface{}
+	Content   any
 }
 
 // kind is not printed right now, because it is added when calling p.publish
@@ -35,10 +37,40 @@ func (e *ISCEvent) String() string {
 	return fmt.Sprintf("%s | %s (%s): %v", e.ChainID, issuerStr, e.Kind, e.Content)
 }
 
+func (p *Publisher) HasSubscribedToAllChains(client *websockethub.Client) bool {
+	return p.subscriptionManager.ClientSubscribedToTopic(client.ID(), "chains")
+}
+
+func (p *Publisher) HasSubscribedToSingleChain(client *websockethub.Client, chainID string) bool {
+	return p.subscriptionManager.ClientSubscribedToTopic(client.ID(), fmt.Sprintf("chains/%s", chainID))
+}
+
+func (p *Publisher) IsClientAllowed(client *websockethub.Client, chainID string, messageType string) bool {
+	if !p.msgTypes[messageType] {
+		return false
+	}
+
+	if !p.HasSubscribedToAllChains(client) && !p.HasSubscribedToSingleChain(client, chainID) {
+		return false
+	}
+
+	if !p.subscriptionManager.ClientSubscribedToTopic(client.ID(), messageType) {
+		return false
+	}
+
+	return true
+}
+
+func BlockEventHandler(client *websockethub.Client) *event.Closure[*publisher.BlockApplied] {
+	return event.NewClosure(func(event *publisher.BlockApplied) {
+
+	})
+}
+
 // PublishBlockEvents extracts the events from a block, its returns a chan of ISCEvents so they can be filtered
-func PublishBlockEvents(blockApplied *publisherBlockApplied, publish func(*ISCEvent), log *logger.Logger) {
-	block := blockApplied.block
-	chainID := blockApplied.chainID
+func PublishBlockEvents(blockApplied *publisher.BlockApplied, client *websockethub.Client, log *logger.Logger) {
+	block := blockApplied.Block
+	chainID := blockApplied.ChainID
 	//
 	// Publish notifications about the state change (new block).
 	blockIndex := block.StateIndex()
@@ -47,7 +79,8 @@ func PublishBlockEvents(blockApplied *publisherBlockApplied, publish func(*ISCEv
 	if err != nil {
 		log.Errorf("unable to get blockInfo for blockIndex %d: %w", blockIndex, err)
 	}
-	publish(&ISCEvent{
+
+	client.Send(client.Context(), ISCEvent{
 		Kind:   ISCEventKindNewBlock,
 		Issuer: "",
 		// TODO the L1 commitment will be nil (on the blocklog), but at this point the L1 commitment has already been calculated, so we could potentially add it to blockInfo
@@ -63,7 +96,7 @@ func PublishBlockEvents(blockApplied *publisherBlockApplied, publish func(*ISCEv
 		log.Errorf("unable to get receipts from a block: %w", err)
 	} else {
 		for _, receipt := range receipts {
-			publish(&ISCEvent{
+			client.Send(client.Context(), ISCEvent{
 				Kind:      ISCEventKindReceipt,
 				Issuer:    receipt.Request.SenderAccount().String(),
 				Content:   receipt,
@@ -79,7 +112,7 @@ func PublishBlockEvents(blockApplied *publisherBlockApplied, publish func(*ISCEv
 	if err != nil {
 		log.Errorf("unable to get events from a block: %w", err)
 	} else {
-		publish(&ISCEvent{
+		client.Send(client.Context(), ISCEvent{
 			Kind: ISCEventKindSmartContract,
 			// TODO should be the contract Hname, but right now events are just stored as strings.
 			// must be refactored so its possible to filter by "events from a contract"
